Added summit xl navigation config files
This configuration has been added here due to it is possible to use the same configuration, however the footprint is different, that's why it is needed to use this package. This package can be used as a configuration package of
summit_xl_navigation by using `ROBOT_NAVIGATION_CONFIG_PACKAGE` environment variable.